miri-unleash tests: ensure they fire even with 'allow(const_err)'
