+++ src/test/ui/consts/miri_unleashed/mutable_const.stderr
@@ -19,7 +19,7 @@ LL | | };
 note: the lint level is defined here
   --> $DIR/mutable_const.rs:5:9
    |
-LL | #![deny(const_err)]
+LL | #![deny(const_err)] // FIXME: ICEs with allow! See #71316.
    |         ^^^^^^^^^
 
 error: aborting due to previous error; 1 warning emitted
